+++ lib/cv_audio.dart
@@ -2,9 +2,7 @@ export 'src/cv_audio_player.dart';
 export 'src/cv_audio_recorder.dart';
 export 'src/cv_audio_router.dart';
 export 'src/cv_audio_transcription.dart';
-import 'src/cv_audio_context.dart';
 export 'src/cv_audio_context.dart';
+export 'src/model/model.dart';
 
-class CarbonVoiceAudio {
-  static CarbonVoiceAudioContext createContext() => CarbonVoiceAudioContext.instance();
-}
+class CarbonVoiceAudio {}