+++ lib/src/cv_state.dart
@@ -1,3 +1,0 @@
-import 'package:carbonvoice_audio/cv_audio.dart';
-
-abstract class CarbonState<T extends CarbonVoiceAudioContext> {}