+++ lib/src/model/model.dart
@@ -1,0 +1,5 @@
+export 'src/audio_file.dart';
+export 'src/audio_input.dart';
+export 'src/audio_state.dart';
+export 'src/recorder_ui_state.dart';
+export 'src/recorder_state.dart';