+++ lib/src/model/src/audio_input.dart
@@ -2,7 +2,7 @@ class AudioInput {
   String? name;
   String? port;
 
-  static AudioInput get none => AudioInput(name: "None", port: "none");
+  static AudioInput get none => AudioInput(name: "Unknown", port: "Unknown");
 
   AudioInput({this.name, this.port});
 }