+++ lib/src/model/src/recorder_state.dart
@@ -1,0 +1,6 @@
+class RecorderState {
+  RecorderState({this.isRecording = false});
+
+  /// audioState
+  bool isRecording;
+}