+++ lib/src/utils/src/permissions/cv_audio_permissions.dart
@@ -1,6 +1,6 @@
 import 'package:permission_handler/permission_handler.dart';
 
-import 'cv_audio_permissions_state.dart';
+import 'cv_audio_permissions_state.dart' if (dart.library.html) 'cv_audio_permissions_state_web.dart';
 
 class CarbonVoiceAudioPermissions {
   PermissionState microphone;
@@ -8,7 +8,7 @@ class CarbonVoiceAudioPermissions {
   CarbonVoiceAudioPermissions({required this.microphone});
 
   static Future<CarbonVoiceAudioPermissions> check() async {
-    return CarbonVoiceAudioPermissions(microphone: await _Permissions.microphone.request());
+    return CarbonVoiceAudioPermissions(microphone: await _Permissions.microphone.check());
   }
 }