+++ lib/src/utils/utils.dart
@@ -1,0 +1,10 @@
+/// Basic utils
+export 'src/audio.dart';
+export 'src/encoders.dart';
+export 'src/file_utils.dart' if (dart.library.html) 'src/file_web_utils.dart';
+export 'src/web_utils.dart';
+
+/// Permission utils
+export 'src/permissions/cv_audio_permissions.dart';
+export 'src/permissions/cv_audio_permissions_state.dart'
+    if (dart.library.html) 'src/permissions/cv_audio_permissions_state_web.dart';